move `DownloadedFile` to submodule
